Remove test_api_edge_cases.py and add process_waiting tests
